initial sysmon-subgraph-generator refactor
