+++ src/rust/grapl-observe/src/lib.rs
@@ -2,3 +2,4 @@ mod metric_error;
 pub mod metric_reporter;
 mod statsd_formatter;
 mod writer_wrapper;
+mod macros;
