+++ src/rust/sysmon-subgraph-generator/src/models/file/mod.rs
@@ -1,0 +1,3 @@
+mod create;
+
+pub use create::generate_file_create_subgraph;
