+++ src/rust/sysmon-subgraph-generator/src/models/network/mod.rs
@@ -1,0 +1,5 @@
+mod outbound;
+mod inbound;
+
+pub use outbound::generate_outbound_connection_subgraph;
+pub use inbound::generate_inbound_connection_subgraph;
