+++ src/rust/sysmon-subgraph-generator/src/models/process/mod.rs
@@ -1,0 +1,3 @@
+mod create;
+
+pub use create::generate_process_create_subgraph;
